RSVPs, styling improvements, glob-http (#89)
* EScape integration WIP 1

* WIP 2

* Added rsvps and a glob-http entry

* fixing make-app.sh

* Fixing script some more

* remove sur file-server.hoon

* Update README

+++ ucal-js/src/views/EventView.tsx
@@ -19,6 +19,7 @@ import { Location, LocationState } from "history"
 import Event, {
   EventForm,
   EventLoc,
+  InviteChange,
   RepeatInterval,
   Weekday,
   WEEKDAYS,
@@ -26,9 +27,9 @@ import Event, {
 import DatePicker from "../components/lib/DatePicker"
 import TimePicker from "../components/lib/TimePicker"
 import { getDefaultStartEndTimes } from "../lib/dates"
-import { addOrRemove } from "../lib/arrays"
-import UrbitApi from "../logic/api"
+import { addOrRemove, findAdditions, findRemovals } from "../lib/arrays"
 import { useCalendarsAndEvents } from "../hooks/useCalendarsAndEvents"
+import { formatShip } from "../lib/format"
 
 const REPEAT_INTERVALS = [
   RepeatInterval.doesNotRepeat,
@@ -41,6 +42,7 @@ const REPEAT_INTERVALS = [
 enum EventField {
   title = "title",
   location = "location",
+  invite = "invite",
 }
 
 interface RouterProps {
@@ -67,15 +69,15 @@ export interface EventViewState {
   startTime: string
   endTime: string
   allDay: boolean
+  invited: string[]
+  invite: string
   //TODO: add all event props here
   event?: Event
   prevPath?: Location<LocationState>
+  inviteChanges?: InviteChange[]
 }
 
 const EventView: React.FC<Props> = ({ location, match }) => {
-  const {
-    authTokens: { ship },
-  } = new UrbitApi()
   const history = useHistory()
   const { calendars, saveEvent, deleteEvent, getEvents } =
     useCalendarsAndEvents()
@@ -87,7 +89,7 @@ const EventView: React.FC<Props> = ({ location, match }) => {
 
   const getCalendarCode = () =>
     (
-      calendars.find((c) => c.title === "default" && c.owner === ship) ||
+      calendars.find((c) => c.title === "default" && c.owner === window.ship) ||
       calendars[0]
     )?.calendarCode
 
@@ -103,7 +105,7 @@ const EventView: React.FC<Props> = ({ location, match }) => {
 
     return {
       calendarCode: getCalendarCode(),
-      organizer: ship,
+      organizer: window.ship,
       title: "",
       desc: "",
       location: new EventLoc({ address: "" }),
@@ -114,17 +116,35 @@ const EventView: React.FC<Props> = ({ location, match }) => {
       allDay: false,
       startTime,
       endTime,
+      invited: [],
+      invite: "",
     }
   }
 
   const [eventState, setEventState] = useState<EventViewState>({
     ...initState(eventToEdit),
-    ...(eventToEdit || {}),
     weekdays: [moment(start).format("ddd").toLowerCase() as Weekday],
   })
 
   const saveEventHandler = async () => {
-    const eventToSave = new EventForm(eventState)
+    const updated = Boolean(eventState.event)
+
+    const eventToSave = new EventForm({
+      ...eventState,
+      inviteChanges: updated
+        ? [
+            ...findAdditions(
+              eventState?.event?.invitedShips() || [],
+              eventState.invited
+            ).map((who) => ({ who, invite: true })),
+            ...findRemovals(
+              eventState?.event?.invitedShips() || [],
+              eventState.invited
+            ).map((who) => ({ who, invite: false })),
+          ]
+        : undefined,
+    })
+
     try {
       const updated = Boolean(eventState.event)
       saveEvent(eventToSave, updated)
@@ -132,15 +152,18 @@ const EventView: React.FC<Props> = ({ location, match }) => {
       await getEvents()
       history.goBack()
     } catch (e) {
-      console.log("SAVE EVENT ERROR:", e)
+      console.error("SAVE EVENT ERROR:", e)
     }
   }
 
   const deleteEventHandler = async (): Promise<void> => {
-    if (confirm("Are you sure you want to delete this event?")) {
+    if (
+      eventState.event &&
+      confirm("Are you sure you want to delete this event?")
+    ) {
       deleteEvent(eventState.event)
       await getEvents()
-      history.replace(eventState.prevPath || "/~calendar")
+      history.replace(eventState.prevPath || "/")
     }
   }
 
@@ -214,6 +237,40 @@ const EventView: React.FC<Props> = ({ location, match }) => {
   }
   const saveDisabled = disableSave()
 
+  const addInvite = () => {
+    const { invite, invited } = eventState
+    const cleanedShip = invite.trim()
+
+    if (cleanedShip) {
+      const formattedShip = formatShip(cleanedShip)
+
+      if (
+        !invited.includes(formattedShip) &&
+        !(formattedShip === `~${window.ship}`)
+      ) {
+        setEventState({
+          ...eventState,
+          invited: invited.concat(formattedShip),
+          invite: "",
+        })
+      } else {
+        setEventState({ ...eventState, invite: "" })
+      }
+    }
+  }
+
+  const removeInvite = (ship) => () =>
+    setEventState({
+      ...eventState,
+      invited: addOrRemove(eventState.invited, ship),
+    })
+
+  const checkEnter = (e: React.KeyboardEvent) => {
+    if (e.key === "Enter") {
+      addInvite()
+    }
+  }
+
   return (
     <Box
       height="100%"
@@ -276,7 +333,7 @@ const EventView: React.FC<Props> = ({ location, match }) => {
             selectTime={selectTime(true)}
           />
         )}
-        <Text fontSize="1" margin="28px 12px 0px">
+        <Text fontSize="1" margin="24px 12px 0px">
           to
         </Text>
         <DatePicker
@@ -340,6 +397,48 @@ const EventView: React.FC<Props> = ({ location, match }) => {
           </Box>
         )}
       </Box>
+
+      <Text fontSize="1" marginTop="20px">
+        Invites
+      </Text>
+      <Row marginTop={eventState.invited.length && "12px"}>
+        {eventState.invited.map((ship, ind) => (
+          <Text
+            className="invite"
+            key={`ship-${ship}`}
+            onClick={removeInvite(ship)}
+          >
+            {ind > 0 ? `, ${ship}` : ship}
+          </Text>
+        ))}
+      </Row>
+      <StatelessTextInput
+        onKeyPress={checkEnter}
+        fontSize="14px"
+        placeholder="Type ship name here"
+        width="30%"
+        margin="20px 0px 0px"
+        onChange={(e) => updateValueHandler(e, EventField.invite)}
+        value={eventState.invite}
+      />
+      <Button width="120px" marginTop="8px" onClick={addInvite}>
+        Add Invite
+      </Button>
+
+      {!!eventState.event?.invited?.length && (
+        <Box marginTop="20px">
+          <Text fontSize="1">RSVPs</Text>
+          <Box>
+            {eventState.event.invited.map(({ ship, status }, ind) => (
+              <Row key={`rsvp-${ship}`} marginTop="8px">
+                <Text>
+                  {ship} : {status || "no response"}
+                </Text>
+              </Row>
+            ))}
+          </Box>
+        </Box>
+      )}
 
       <StatelessTextInput
         fontSize="14px"
